% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poem_report.R
\name{poem_report}
\alias{poem_report}
\title{Generate template for creating poem report}
\usage{
poem_report(report_month, report_year)
}
\arguments{
\item{report_month}{Month of report}

\item{report_year}{Year of report}
}
\description{
Opens a template of the poem report for the month, optionally rendering it.
}
